// repos/tvl-labs/khalani-ui/src/components/TableArchived/Table.types.ts
import { ReactElement } from 'react'

import { TableProps } from '@mui/material/Table'
import { TableCellProps } from '@mui/material/TableCell'

type fieldValue = string | number | ReactElement

export interface IColumn {
  value: fieldValue
  name: string
  width: string
  align?: TableCellProps['align']
  sortOrder?: ESortOrder
  sortDefault?: boolean
}

export interface IRow {
  id: string
  cells: {
    [key: string]: {
      value: number | string | ReactElement
      sortingValue?: string | number
      filteringValue?: string[]
    }
  }
}

export interface TableHeadProps {
  columns: IColumn[]
  onSortColumn?: (value: string, sortOrder: ESortOrder) => void
  displayHeader?: boolean
}

export interface TableBodyProps {
  columns: IColumn[]
  rows?: IRow[]
  RouterLink?: React.ElementType
  redirectPath?: string
  isLoading?: boolean
  handleRowClick?: (id: string) => void
}

export interface ITableProps
  extends TableProps,
    TableHeadProps,
    TableBodyProps {
  filteringParams?: FilteringParams
}

export interface FilteringParams {
  tokenSymbol?: string
  network?: string
}

export enum ESortOrder {
  ASC = 'asc',
  DESC = 'desc',
}

export enum EAlign {
  RIGHT = 'right',
  LEFT = 'left',
}